Conflicts:
	app/views/backend/parcels/index.html.haml
	config/locales/eng/procedures.yml
	config/locales/fra/enumerize.yml

+++ app/models/activity_production.rb
@@ -192,10 +192,7 @@ class ActivityProduction < Ekylibre::Record::Base
   end
 
   after_destroy do
-
-    if support.is_a?(LandParcel)
-      support.destroy
-    end
+    support.destroy if support.is_a?(LandParcel)
 
     Ekylibre::Hook.publish(:activity_production_destroy, activity_production_id: id)
   end
@@ -224,10 +221,10 @@ class ActivityProduction < Ekylibre::Record::Base
   end
 
   def initialize_land_parcel_support!
-    support_shape ||= cultivable_zone.shape if cultivable_zone
+    self.support_shape ||= cultivable_zone.shape if cultivable_zone
     unless support
-      if support_shape
-        land_parcels = LandParcel.shape_matching(support_shape)
+      if self.support_shape
+        land_parcels = LandParcel.shape_matching(self.support_shape)
                                  .where.not(id: ActivityProduction.select(:support_id))
                                  .order(:id)
         self.support = land_parcels.first if land_parcels.any?
@@ -235,14 +232,14 @@ class ActivityProduction < Ekylibre::Record::Base
       self.support ||= LandParcel.new
     end
     support.name = computed_support_name
-    support.initial_shape = support_shape
+    support.initial_shape = self.support_shape
     support.initial_born_at = started_on
     support.initial_dead_at = stopped_on
     support.variant ||= ProductNatureVariant.import_from_nomenclature(:land_parcel)
     support.save!
     reading = support.first_reading(:shape)
     if reading
-      reading.value = support_shape
+      reading.value = self.support_shape
       reading.save!
     end
     self.size = support_shape_area.in(size_unit_name)
@@ -579,9 +576,7 @@ class ActivityProduction < Ekylibre::Record::Base
   def current_cultivation
     # get the first object with variety 'plant', availables
     if cultivation = support.contents.where(type: Plant).of_variety(variant.variety).availables.reorder(:born_at).first
-      return cultivation
-    else
-      return nil
+      cultivation
     end
   end
